Retrieve products in `products FC` with `useStaticQuery`
I didn't read the Gatsby docs well as to the difference between
types of queries depending on what sort of file it is in.

Page queries only work in pages, like `index FC`, but not in
`products FC`. Which is why I spent a lot of time having the exact
same query moved to the products page fail. Now it works as intended.

Addresses #1 (Pull anything from GraphCMS)
Project: 'Integrate GraphCMS'